Pre-test hook for integration tests/deprecated
We want to allow for integration tests to test deprecated panels.  This
means that we need to override the configurations provided by the panels
and force a 'legacy' setting.

On the Infra side, we need to pass in an environment variable that tells this
override to be enabled.

Depends-On: I20232ef16ae60a7151e267b506af471ebfed2f7f
Change-Id: If5a3e273098aa726e12f9076c0273a5afaf233d2
Partially-Implements: blueprint test-deprecated-features