allow hiding help-button for first ng workflow step
we have merged patch that will hide the help-button on
workflow steps if helpUrl is not defined
https://review.openstack.org/#/c/299705/

However, it forgot to factor in the 'first' step.
With this patch, you can see it used on its parent patch:
Create Volume action removed the helpUrl and the
helpBtn doesn't show up either.

Change-Id: I7e3ee4d03b9f5f1fa924d7f233e4fa4ea4ed12a7