Remove all_tenants flag in quota check
I don't really know what this was meant to do, but what it actually
does doesn't make any sense -- when checking a quota for project
'foo' we really don't care about how many instances are used
in the entire cloud, only in project 'foo'.

Until Liberty, the all_tenant flag was broken in the nova
api, which probably explains why this was harmless until
Liberty.

Change-Id: I9e12c10951e7518ab1c622a4daac962bad332da5
Closes-Bug: #1610693