Let Nova to pick availability zone if more than 1
In the Angular Launch Instance, if there is more than one availability zone,
default to the option for the Nova scheduler to pick.  This is a regression
from the legacy Launch Instance feature.

If you want to simulate testing the logic with multiple avaiability zones,
go to:  .../launch-instance/launch-instance-model.service.js line 785 and add:

model.availabilityZones.push({label: 'another one', value: 'another one'});

Change-Id: Ib81447382bc9d43e33ce97f78c085d2a94ff2018
Closes-Bug: 1613900