Turn on docs warning check in document generation
Use -W (turn warnings into errors) option of sphinx-build
in the commandline of 'docs' tox target so that developer
can easily check sphinx warning.

Also runs the same documentation check in 'pep8' tox target
to detect sphinx warning in the gate.
The current 'docs' job in the gate does not use 'tox -edocs'
intentionally and calls build_sphinx via 'tox -evenv' [1],
so sphinx warnings are not detected in the 'docs' job.

Note that we no longer generate the whole code reference so this
change does not increase the time of 'tox -epep8' much,
while we can prevent new sphinx warnings.

[1] https://github.com/openstack-infra/project-config/blob/6b50d7e3a69c49d1e035c6d026d8e3910e62a981/jenkins/scripts/run-docs.sh#L16-L19

Closes-Bug: #1411719
Closes-Bug: #1486222
Change-Id: Idc6e8a1c5762eba113b2d110d5fa223ab7406be3