A following up patch for 'Added the pagination info at top of the table'
Originally, 'Displaying x items' is included in footer area.
However, the element that the previous patch added is not in this area.
Thus, when the number of rows in table is zero, this behaviour is difference
(the element at the bottom of the table is hide, but the element
at the top of the table remain displaying).
I think these should be same behaviour as a footer information.

Change-Id: I6eb10db2bf00c8686c0df9d990787a2a890e6340
Related-Bug: #1359238