Remove unimplemented policy checks for Neutron
The dashboard is doing policy checks for Neutron that aren't even being
done by Neutron itself.

This fixes bug #1628141, which was caused by change-id
I9f4e9209606999e5529e5ba068640d607b817f56, which was meant for Nova and
not for Neutron.

This change makes the policy checks for Nova secgroups only, if Neutron
networking is not enabled.

Neutron policy checks removed:

 * delete_security_group
 * create_security_group
 * update_security_group
 * get_security_group
 * create_security_group_rule
 * delete_security_group_rule

Change-Id: I46b46fcd4cbc7c8e06f481eac9606c330fc75351
Closes-Bug: #1628141