Remove hardcode of container_name in backups
If user doesn't set container name
for volume backup, Horizon sets default
value 'volumebackups'. If Cinder is configured
with different default value, this behaviour
will cause an error.

Since Cinder is able to handle case when
container name for backup is not set,
we should remove this hardcode from Horizon.

Change-Id: Ib121ac1d38f11e967ef059de57a8198c5e890198
Closes-Bug: #1631930