Link to the user details from action log
The "User ID" column in the action log is not very usable, showing
an unreadable ID of the user, instead of the user name. However,
getting the name of the user for every row could prove expensive.
Therefore, I'm turning the ID into a link to the user details page,
so that it's easier to see what user it means.

Change-Id: Ic07903a145ee4a3decc71f12717ed06e84fed0eb