Fix the issue that a wrong message is shown in ng-launch instance
In some case like an api response is delaying, users may see the message
"There are no Availability Zones" in Details tab.
Because the request to get availability zones is not still finished.
It makes user confuse.
This patch distinguish whether there are no availability zones or just
under requesting it and show users a proper message.

This matter may occur not only above case but also other tabs.
But at the moment, this patch is addressing only Details tab because
this tab is first one to be shown to users.

Change-Id: Iec1c18ced37b0c3a5f25a01975a8e2b0f2c1f335
Closes-Bug: #1642108