Ensures proper network name in subnet error message
We need to pass network object to _create_subnet()
to show proper network name in subnet error message.
The network itself has been retrieved in the view code, so what
we need to do is to make it available in the workflow code.
This commit makes 'network' information available in the workflow
context_seed dict and CreateSubnet handle() method can use it.

Also removes the unnecessary network_get in the admin subnet create
form. we can retrieve a network object from context_seed['network']
as we do in the project side code now.

Change-Id: Ib2162814cbc7263fc03a9da0319c0b2b38732cc5
Closes-bug: #1505414
Co-Authored-By: Akihiro Motoki <amotoki@gmail.com>