Add constraints support to run_tests.sh
Currently developers need to manually build a constratined venv if they
want to  match the versions that are tested in the gate.  This change
adds logic to run_tests.sh and install_venv*.py to default to the same
constraints file used in the gate.

With constraints enabled the 'churn' in the environment should be
minimised so remove the developer optimisation for detecting when
packaged versions change and always just install/update.  This has the
down side of making the script preamble slightly longer but we always
print_help() so I don't think it'll be terribly inconvenient.

Change-Id: I9bf85adb5aa2c85d8c0f0f22ac9efe5cac3646c0
Closes-Bug: 1644697