Enable sorting by size in containers pane
The containers module was using an incorrect ID in the "Size" column,
preventing the user from sorting by "Size". This change updates the
ID s/size/bytes/ which is the resource property holding the actual object size
in Bytes.

This change also updates the default "subdir" size to 0, that is the
default value returned in the "Content-Length" header for "subdirectories"

Change-Id: I98d164ec1a08d15f25fcdbc92f45e46c9fa6e5c4
Closes-bug: 1643710