In drawer, handle 'priority' attribute conversely against row
'priority' attribute has be handled in drawer [1][2], but the behavior
is same as in rows. The behavior is that the items which have lower
priority are hidden from both of row and drawer when wid$

Drawer is used for the supplements against the row, so it should be
handled conversely.
It means that items appear in drawer when they are hidden from row.

This patch fixes it.

[1] https://bugs.launchpad.net/horizon/+bug/1648332
[2] https://review.openstack.org/#/c/397132/

Change-Id: Ia0487a1d95ba16d764546cdb374fe786a0f0f648
Closes-Bug: #1651658