Uniform the button name and urls to create subnet.
For the button's name, it is sometimes named as 'Add Subnet' and
sometimes named as 'Create Subnet'. However, they both direct
to the same 'create_subnet' API/function. To avoid the confusion
of these two names, 'Add Subnet' is updated to 'Create Subnet'.
Accordingly, the last part of urls are updated to createsubnet.

Change-Id: I8e3198c0f3945da5fd46aade60429f3638d73523