Fix single policy rule handling
With commit 43e9df85ab286ddee96e9cff97f551781baf70d1 the handling
of single policy rules was broken and always returned True for a
single rule.

One of the visible results is that the Admin Dashboard showed up
incorrectly for users that lacked permission to see it. Additionally,
panel in the Admin Dashboard were also visible.

This patch fixes single rule handling, and the visible effects.

Closes-Bug: #1653792
Change-Id: I0c8a0d7b230b6c6b7ee048af84646ca95daee340