Check if user.token attribute exists
If the user is admin or don't have an ID the identity panel
will be loaded. In any other case  the project panel will be loaded.
This will also check if the panel is unregistred, and if the
session has expired it will lead to the login screen.

Closes-Bug: #1659818
Change-Id: I27300e928855ec10ca5d2a623d636e50d756b698
(cherry picked from commit a2386d80fbbff2cd3ab3814cd850a8471280dde8)