Pass the project domain ID to novaclient
When the nova client creates a new session, knowing only the ID and
project of a token but not the project's domain, it calls keystoneauth
with None as the project domain. Keystoneauth takes this to mean
novaclient wants to authenticate with the Identity v2.0 API, which
fails if the user is a federated user. This patch grabs the project
domain ID from the request and passes it along to novaclient so that it
can properly create a v3 Token session.

Partial-bug: #1660436
(cherry picked from commit f7e2d7883dc652085b2982c3a3ea766cb7993c68)

Change-Id: Iaf331ecd7b02076be6f9123baf59de787cdbc3db