Make VPN IPSec Site Connection PSK field hidden
The Pre-Shared Key (PSK) field on the VPN IPSec Site Connection tab
should not be displayed in plain text due to security concerns. Set
the PSK field in the Add Connection and the Edit Connection dialogs
to be a password field to provide the user some protection when
entering the value. Remove the PSK field from the details page since
this is the pattern used with the password field in Identity Users
panel.

Change-Id: I4dd713f01b02c29d9822efcb519de60fd9d035e6
Close-Bug: #1575909
(cherry picked from commit 5137dc4fdd19de3494293731abffdfb7e5b26449)