Add rule "update_network:shared" to neutron_policy.json
This rule is referenced by networks panel and is required for
enabling non-admin users to change the name of networks.

This change has been committed to master in commit
d599fdec599db99f3c8b73ffce18a140bae8e629

That commit is too big for back-porting to stable/mitaka as-is,
thus I am proposing this one-line commit to fix the issue at
hand for stable/mitaka.

Change-Id: I93ef6e09f37557dc2375d7672d2a247c776ec2bb
Closes-Bug: #1609467