Use BooleanField for admin_state_up form
admin_state_up is a boolean value but the network and port edit forms
expected a string version of True/False. As a result, True/False string
was shown as the default value in these forms. The field is implemented
as ChoiceField, but there is no special reason not to use BooleanField
for a boolean field and admin_state(_up) fields are the only exceptions
in horizon. This commit replaces all admin_state(_up) fields to
use BooleanField.

As far as I checked, this pattern is used only in the networking
related panels and this patch clean them up.

Conflicts:
	openstack_dashboard/dashboards/project/networks/ports/forms.py
	openstack_dashboard/dashboards/project/networks/workflows.py

Closes-Bug: #1672213
Change-Id: I9286f6c29d67fec7a88b74661bc8eca945fd9061
(cherry picked from commit c8cd03189a7ea38ebdd97d613a88a50d255005c2)