Do not try to show L3 agent action when enable_router is False
When enable_router settings is False, 'routers' panel is not
registered and URL resolver of L3 agent action in the network
agent table will result in a failure.
As a result, System Information panel cannot be displayed.

This commit adds a check whether the router panel is enabled or not.

Change-Id: I67aba79f61870b093d4c1d9c7527003704483d50
Closes-Bug: #1671785
(cherry picked from commit f54c52418bdae0403353447d688b77be15e8a044)