Option to specify MAC address of port
Enable display and manipulation of the MAC address of a port.

That is:
(1) Display MAC addresses in the ports tab of the networks panel.
    Note: Port detail shows the MAC address with or without this patch.
(2) Allow input of MAC address at port create in both project and admin
    dashboards.
(3) Allow input of MAC address at port update in admin dashboard only
    because by default MAC update is an admin-only operation in neutron
    (though this can be changed by policy configuration).

Change-Id: I60fb4921f08286d90469cd420261c7e600f5bc7c
Partially-Implements: blueprint neutron-trunk-ui
Closes-Bug: #1667778