table action: drop deprecated action_present/past attributes
action_present and action_past "attributes" have been deprecated
since the middle of Mitaka development cycle. There is no consumer
in the horizon repository and the deprecation warnings were emitted
appropriately so horizon plugins are expected not to use them already.
Now we are in the early stage for the next release (Pike),
so it is time to drop the deprecated way.

Change-Id: Ia93a6b17561068777845366b2b4c57b7144fab9f