Enable warning-on-error in Sphinx 1.5
warning-on-error option is avaialble in sphinx setup command integration
since Sphinx 1.5 [0]. This is useful to prevent future warnings.
Note that there is no sphinx error or warning in the current
horizon code base :)

Also renames 'all_files' to 'all-files' in [build_sphinx] section
as this is the actual option name defined in sphinx [0].
The underscore version 'all_files' also works as pbr
handles it properly, but we use hyphen version for sphinx options,
so the consistent usage is less confusing.

[0] https://github.com/sphinx-doc/sphinx/blob/1.5.3/sphinx/setup_command.py#L69

Change-Id: Ic9d1b0e4db30891c694537d43d7f7a51ff92be49