Save image name in Server.image_name method
The api.nova.Server.image_name @property can be called multiple times
for a given server. It can result in so many calls to glance.get_image.

This patch saves the result of the image name as returned by the
glance.get_image call in the Server.image dict, so that subsequent calls
to image_name can return faster.

Change-Id: I38993f13619f1725bcc7c763fe364f415e88eb62
Closes-bug: #1676420