Set minimum volume size based on the flavor selected
On the Angular launch instance modal, the volume size for the new
instance does not take the disk size required by the selected flavor
into consideration, which is inconsistent with the Django launch
instance modal.

This patch will set the volume size of the new instance to the
minimum disk size required by the selected flavor.

Change-Id: Ifb35b70c5b41820024e17da783687bfe7880b8e8
Closes-bug: #1676648