Move all RC download buttons under a single menu
In the API access page, we now see three or four
"Download OpenStack RC" buttons.
This commit packs them into a single menu.

To complete the goal, a new option called "table_actions_menu_label"
is newly introduced to DataTable.

Also the condition to disable an action dropdown menu is changed
so that the action dropdown is not disabled when "multi_select"
feature is not used in a corresponding table.
Otherwise, the dropdown menu is always disabled and there is
no way to enable it in a DataTable with multi_select=False.

Change-Id: I229c69b09c45bb20b3df48c9901d76b89fd27ee4
Closes-Bug: #1680563