Include logging level to log message by default
Currently logging level is not included in the default configuration.
Even though it can be customized, it would be nice and helpful
if the default logging configuration includes it.

Apache2 mod_wsgi outputs timestamp when logging messages,
so I do not include %(asctime)s in the proposed format and
removes it from the operation log format.
If WSGI framework you use does not output timestamp you can add it
to the formats and it is mentioned as the inline comment.

Closes-Bug: #1681507
Change-Id: Id0efdd00712ac71ef1739364e441918d4648d88c