Fix a message bug after a failing live migrate
If live migrate operation failed, there would be a message like:
'Failed to live migrate instance to host "host01".'
But, if we choose none target host, the message will be:
'Failed to live migrate instance to host "".'
The empty quote marks looks strange.

I think it would be better to check whether data['host'] is
None/empty at first and use a message without hostname if it is.

Change-Id: Ib25f2d9481cfe9a5119d3473ba7ea8032720f45a
Closes-Bug: #1684394