Check the target host before live migrating a instance
In the live migrate form, the selections "Automatically schedule
new host." and "No other hosts available" both use an empty value,
and this input is not required, that causes I can submit the
form even there are no other hosts available.

I think the value of selection "Automatically schedule new host."
should not be empty, and this input should be required.

Change-Id: I6a88ffa23087a0d845cf8a71a6359f3f4fbddbe0
Closes-Bug: #1684604