Allow Chrome to start when tox is running npm
Chrome in *nix environments requires the HOME and DISPLAY variables
to be set for Chrome to start properlyi when running the 'npm' rule.
The HOME variable is needed so Chrome can load up the user's
preferences.  The DISPLAY variable is needed so it know what display
to launch Chrome into.

Change-Id: I84f4d1d47f065158e8f5bf5691c241061cc16aab
Closes-Bug: #1684321