Generate unit test HTML reports
Currently we need to check unit test results by opening console.html.
HTML reports would be nice for this purpose.

Note that the infra CI post script stores '*nose_results.html',
so the filenames need to end with 'nose_results.html'.

--with-html-output option is not available for python 3. I am not sure
the reason at now, but I think it is worth enabling it for python 2.

Change-Id: Icea8d21d0cf93f99f577e6609e9ced166a55609a