Fix a filter error in images panel.
In images panel, I set the filter to be "visibility: unknown",
then an error appeared at top right corner of the page:
"Error: Unable to retrieve the images."

I open the developer tool of browser, I found an error response
in the network tab:
"Invalid visibility value: unknown".

The possible values for image visibility includes public, private,
shared and community, so I fixed it in the filter.

Change-Id: I48fc32ed37ac1c8ef36d0a0b8c05082c649a8534
Closes-Bug: #1691669