Attribute "isServer" of images visibility filter should be true.
I found an interesting bug in images panel.
In images panel, set filter to be "visibility: public", I got one
image listed in the table, then I changed language from "en" to
"zh_CN" and filter images again, I got none.

In the tab of browser developer tool, I saw a correct response
with one image data, but it was not appeared in the table. Later
I found the attribute "isServer" of visibility filter is false,
that caused the above bug.

Change-Id: I130d07727fe74ce65aba72d3e63fdfe0c8721795
Closes-Bug: #1691683