Enable to refresh ngdetails view
For now refreshing ngdetails view by browser using F5 key or reload
button, it causes 404 error from django. To fix this issue, this patch
adds the url for '/ngdetails'.

Furthermore, to specify current navigation and to check access grants
to current panel, each access to Horizon django framework needs its
dashboard and panel objects. It means that we need to specify dashboard
and panel that ngdetails view belongs to.

Also, this patch adds the process and settings to specify dashboard and
panel object for each ngdetails view.

Change-Id: I4bcffd2e222ce2df186551cceba0aa38f600d9c8
Closes-Bug: #1681627