Add swift unit tests to improve coverage
Add jasmine unit tests to cover swift.service.js to increase the
branches coverage.  The coverage threshold for branches in
openstack_dashboard, 93.05%, was barely over the minimum threshold, 93%.
That small margin made it impossible to remove more than 1 line from an
if-statement without dipping below the minimimum threshold and failing
the check job.

Change-Id: I0020ca6d5fdae436153f31badfa2b97bcd45a0ee