doc: better formatting of TOC in sidebar
After switching openstackdocstheme, TOC in the left sidebar becomes long.
This is because openstackdocstheme treats the first level of 'doctree'
in the top page (index.rst). In Horizon documentation, many documents
are listed as the first level, so the sidebar is long.

This commit adds index.rst per topic and merges all toctree
under the developer docs into a single toctree.
This will lead to the well-organized toc tree in the sidebar.
(Note that this does not touch "Using Horizon" part as it is being
refactored in https://review.openstack.org/#/c/473932/.)

The title level of "Packaging Horizon" in topics/packaging.rst is
changed to the second level because openstackdocthemes assume
only one first-level title per file and the second first-level title
or later are not displayed in a rendered page.

Implements: blueprint pike-docs-overhaul
Change-Id: I16e2bae18d8f67716c9838094a39527cd8bf733e