Fix subnets can not be deleted from network topology panel
The structure of DetailView was changed, so the url for the delete
subnet action needs to be updated so that subnets can be deleted properly.

Added the required method get_subnets_data in DetailView.

Added JS code to remove the subnet from the network node on network
topology panel after the subnet is deleted.

Fixed router interface not being removed from the network topology
panel after it's deleted.

Closes-bug: #1643414

Change-Id: I60cc161655b994c8e3ace5efc37459f3e9e60d67