Incorrect URL for polling heat stack
There is a "topology" tab in the Horizon page which shows heat stack details.
It shows stack's resource dependencies graph. Horizon polls for stack changes
every few seconds.

Currently these pool requests fail in some environments with errors in the
apache log. This is due to the WEBROOT not being taken into account which would
change the path required. This patch fixes this issue.

Change-Id: I894d8ab97bc3f67d408ce2f5496da1570d2fb1d3
Closes-Bug: #1707619