do not flush selection on "old" boot source
This fixes an issue introduced by watchers:
when page initialized images watchers called before
select boot source watcher.it leads to sutiation when
we specify selection from user input and flush it on
updateBootSource. as a result page come to
inconsistent state described in the bug.
Perhaps, this requires big refactoring but  in the
current patch we do not flush selection if boot type
was not changed. it allows to use image, snapshot,
volume from context as ctrl.selection.

Change-Id: I4b23071a2a9e2c02cc1f46713165e5cb5155894f
Closes-Bug: #1710890