extract_messages, update_catalog: Simplify help message
Previously the help message of update_catalog shows all available
languages in the command example. It is difficult to understand
what an exact command line is expected.
This commit uses metavar to simplify the brief command line
and shows available languages in the help text.

Similar change is made to 'domain' argument of update_catalog
and of extract_messages.

Also the default values are added in the help message.

Change-Id: I63efea7e368eb1dfaf203ca4c515eb57173b571e