Handle binding__host_id conditionally
binding__host_id (binding:host_id in the Neutron API) only exists
when portbinding neutron extension is available. This extension is
available in almost all cases as Nova depends on the extension,
but it is better to check the availability in horizon side as
we do so in other places in horizon.

Change-Id: Id7fa8ebcdde09547fd64892a787834bc4c543acf
Closes-Bug: #1498352