Show a better error message for network delete failure
Previously deleting a network shows an unclear error message
and it is not easy for a user to identify the reason:

  Error: Failed to delete network <network_name>

With this fix it will now show proper error message including
the error reason when neutron returns 409 (Conflict) response.

NOTE: This is a fix for the https://review.openstack.org/#/c/140614/
patch, which I couldn't restore.

Change-Id: I5779725f1095d59a012d8e6cf6e47bbd11dda36a
Closes-Bug: #1400667