Allow regular users to specify VNIC type for port
VNIC type needs to be specified when creating a server with an SR-IOV
port. In this case, a user first creates a port with vnic_type=direct
and creates a server with the created port. To allow this scenario,
horizon needs to allow users to specify VNIC type for port.

This commit also clean up the code duplication on the port security and
mac state fields in the port create form. The code in the admin form for
them are completely duplicated with that in the project form.

Change-Id: Ib6e91ed7f720e2720994025429da0dcef2fa4e25
Closes-Bug: #1402959