Support simple FIP disassociation (with FIP release)
We previously supported so-called "Simple FIP disassociation"
which allows users to disassociate and release a FIP in a single action.
We no longer support nova-network based features, but I believe it is worth
implemented even in a neutron-only era. This patch introduces a checkbox
"Release floating IP" to support this with neutron.

At the same time, this patch also fixes a bug that the existing FIP
disassociation action disassociates and releases a first FIP of
a requested server. Even though it is a rare case where a single
server has multiple FIPs, this is a bug. After this patch, FIPs
associated with the requested server are listed in the form and
a user can select an FIP to be disassociated.

This patch drops a setting parameter 'simple_ip_management' without
deprecation notice. This is actually no side effect because this setting
just toggled the FIP disassociate action in the instance table and
it provides nothing more than that. We can do the same thing by
the policy file.

Change-Id: Ie8053bdd3a3e4c7897c7c906788d40c2a1d3f708
Closes-Bug: #1226003