Fix incorrect volume type value in ng images create volume form
The value of the 'volume_type' is determined by the 'volumeType'
object [1]. The value of the 'volume_type' changes only the
'volumeType' object changes. However, the current value of the
'volume_type' is changed according to the volume object change [2].

It will cause a phenomenon, when the page is initialized, the value
of 'volume_type' is empty, when we switch volume types drop-down box,
its value is also empty, only change the name, description, size and
availability zone, the value of 'volume_type' is the page selection.

[1] https://github.com/openstack/horizon/blob/master/openstack_dashboard/static/app/core/images/steps/create-volume/create-volume.html#L45
[2] https://github.com/openstack/horizon/blob/master/openstack_dashboard/static/app/core/images/steps/create-volume/create-volume.controller.js#L140

Change-Id: If754d0c2ced844414c35829d4cefa1fb861522d5
Closes-Bug:#1737900
(cherry picked from commit 3aee4cbac666f080133671790baa9dd42223c996)