Fix batch delete key pairs button isn't disabled when the key pair has deleted
When the key pair has deleted by the batch delete button, the batch
delete button isn't disabled and still can be clicked.

Since whatever in the keypairs panel or on key pair details page,
After the key pair is deleted, the url will be redirected to the
key pairs panel. The correct way is that only the key pair is deleted
on the details page, and it needs to be redirected to the key pairs
panel.

Change-Id: Ie3041cc70bde84e9753be41f0a80d04534ffdf44
Closes-Bug:#1739871